pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yolo-pose-app"
version = "0.1.0"
description = "YOLO Pose Detection Web App using Streamlit"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "streamlit>=1.31.0",
    "ultralytics>=8.1.0",
    "opencv-python>=4.8.1",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["app.py"]